refactor: 도메인별 문서 분리 + 타입 파일 분리
- CLAUDE.md 경량화 (210줄 → 120줄): DB 스키마 제거, 도메인 문서 포인터로 교체
- docs/domains/ 생성: schedule.md, routine.md, insight.md, budget.md
- lib/types.ts → 도메인별 types.ts 분리 (schedule, routine)
- 공통 타입(CategoryRow, 색상 시스템)만 lib/types.ts에 유지
- 30개+ 파일 import 경로 업데이트
- 보안 규칙 강화: 개인 재정 정보/상황 공개 금지 조항 추가
