+++ web/src/features/routine/components/routine-page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useCallback } from 'react';
-import type { RoutineRecordRow } from '@/lib/types';
+import type { RoutineRecordRow } from '@/features/routine/lib/types';
 import { useRoutines } from '../hooks/use-routines';
 import { DateNav } from './date-nav';
 import { ViewToggle } from './view-toggle';
