+++ web/src/features/schedule/components/day-detail-panel.tsx
@@ -2,8 +2,9 @@
 
 import { format } from 'date-fns';
 import { ko } from 'date-fns/locale';
-import type { ScheduleRow, CategoryRow } from '@/lib/types';
-import { compareSchedulePriority } from '@/lib/types';
+import type { ScheduleRow } from '@/features/schedule/lib/types';
+import { compareSchedulePriority } from '@/features/schedule/lib/types';
+import type { CategoryRow } from '@/lib/types';
 import { ScheduleCard } from './schedule-card';
 
 interface DayDetailPanelProps {
