+++ web/src/features/schedule/components/week-view.tsx
@@ -4,8 +4,10 @@ import { useRef, useEffect } from 'react';
 import { startOfWeek, addDays, format, isToday } from 'date-fns';
 import { ko } from 'date-fns/locale';
 import { useDraggable } from '@dnd-kit/core';
-import type { ScheduleRow, CategoryRow } from '@/lib/types';
-import { getCategoryStyle, compareSchedulePriority } from '@/lib/types';
+import type { ScheduleRow } from '@/features/schedule/lib/types';
+import { compareSchedulePriority } from '@/features/schedule/lib/types';
+import type { CategoryRow } from '@/lib/types';
+import { getCategoryStyle } from '@/lib/types';
 import { computeWeekLayout, WEEK_START, type WeekSpan } from '@/features/schedule/lib/calendar-utils';
 import { StatusBadge } from './status-badge';
 import { DroppableDay } from './droppable-day';
