+++ web/src/features/schedule/hooks/use-backlog.ts
@@ -1,7 +1,8 @@
 'use client';
 
 import { useState, useEffect, useCallback } from 'react';
-import type { ScheduleRow, CategoryRow } from '@/lib/types';
+import type { ScheduleRow } from '@/features/schedule/lib/types';
+import type { CategoryRow } from '@/lib/types';
 
 export function useBacklog() {
   const [schedules, setSchedules] = useState<ScheduleRow[]>([]);
